Move file handling outside of STLParser.parse function.
I believe this will make testing the class much easier. I won't need to either provide a whole collection of files or fake the open() function that was previously inside the parse call. Plus this gives the caller more freedom. Maybe the mesh isn't in a file on disk but they still want to parse it.