Overhaul to Camera and CameraController dolly.
The command is now called scale. We reserver Camera.dolly to strictly moving the camera along its z. We use dollying for perspective projections and we use zooming for orthographic projections. There is a zoom for perspective projections but there is no need to use it.

We broke the zoom to cursor functionality but we will now work to put that back.