Super WIP: Very very basic mouse picking.
Suffers from a number of problems. Mainly it only does AABB intersection testing and settles for the first AABB it hits. This needs to check against triangles and then sort the distance to the intersection to find the minimum. Also the call stack getting from scene intersection down to triangle intersection is a little bit nuts right now.