Implement more flexible messaging system.
Instead of relying on inheritance, we use decoration to allow objects to emit and listen to messages. This allows classes to be able to handle both emitting and listening simultaneously. It also allows more flexibility in class method naming. No longer does the function name need to match the Event name.