Major WIP: Start untangling the rendering entities.
Instead of making visible things an object, I am thinking of just having a bunch of rendering functions (one for each type of thing rendered) and then mapping the actual simulation objects (e.g. serial) to a rendering function: one to render the robot, one to render the frame, one to render the tool, etc. This should make an eventual GUI to control visibility easy as well.